Replay helpers stage one
This is the first step in implementing so-called "replay helpers". A
replay helper is a second instance of RetroArch which connects via
netplay in order to perform the replays on behalf of its host. This way,
replays can happen in a separate process, sidestepping many of the
issues with replay that arise due to thread timing.

Replay helpers are not yet complete, and at this stage some important
components are not communicated properly.