Streamline replay helpers
Replay helpers no longer need to send every frame, only the frames that
the host should load. This does mean, however, that the host is 100%
dependent upon the replay helper; the host can never perform its own
replays, and the replay helper must be constantly engaged in that
pursuit. This ALMOST brings PSX into the realm of possibility.